rewritten model loading to match new project tree

+++ include/engine/renderers/scene_3d.cpp
@@ -53,7 +53,7 @@ void Scene3D::render() const
 		getShaderProgram()->setUniformVector3("dir_lights[" + i_str + "].color", m_directional_lights[i]->getColor());
 		getShaderProgram()->setUniformVector3("dir_lights[" + i_str + "].direction", m_directional_lights[i]->getDirection());
 	}
-	getShaderProgram()->setUniformUint("num_dir_lights", m_point_lights.size());
+	getShaderProgram()->setUniformUint("num_dir_lights", (unsigned)m_point_lights.size());
 
 	for (unsigned i = 0; i < m_point_lights.size(); i++)
 	{
@@ -62,7 +62,7 @@ void Scene3D::render() const
 		getShaderProgram()->setUniformVector3("point_lights[" + i_str + "].position", m_point_lights[i]->getPosition());
 		getShaderProgram()->setUniformFloat("point_lights[" + i_str + "].range", m_point_lights[i]->getRange());
 	}
-	getShaderProgram()->setUniformUint("num_point_lights", m_point_lights.size());
+	getShaderProgram()->setUniformUint("num_point_lights", (unsigned)m_point_lights.size());
 
 	for (unsigned i = 0; i < m_spot_lights.size(); i++)
 	{
@@ -74,7 +74,7 @@ void Scene3D::render() const
 		getShaderProgram()->setUniformFloat("spot_lights[" + i_str + "].inner_angle", m_spot_lights[i]->getInnerAngle());
 		getShaderProgram()->setUniformFloat("spot_lights[" + i_str + "].outer_angle", m_spot_lights[i]->getOuterAngle());
 	}
-	getShaderProgram()->setUniformUint("num_spot_lights", m_spot_lights.size());
+	getShaderProgram()->setUniformUint("num_spot_lights", (unsigned)m_spot_lights.size());
 	
 	getShaderProgram()->setUniformVector3("camera_position", m_camera_ptr->getPosition());
 	auto projection_view_matrix = m_renderer->getProjectionMatrix() * m_camera_ptr->getViewMatrix();
